# openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: A plugin for ChatGPT that allows users to query threat intelligence
    off OSINT sources.
  servers:
  - http://localhost:9001
  title: ChatGPT OSINT Plugin
  version: 0.1.1
openapi: 3.0.2
paths:
...
  /api/v1/hybridanalysis/analysis-overview:
    post:
      description: Get analysis report of a SHA256 hash against Hybrid Analysis.
      operationId: analysis_overview_api_v1_hybridanalysis_analysis_overview_post
      parameters:
      - in: query
        name: sha256
        required: true
        schema:
          title: Sha256
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Analysis Overview Api V1 Hybridanalysis Analysis Overview
                  Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Analysis Overview
  /api/v1/hybridanalysis/report-summary:
    post:
      description: Get the report summary for a SHA256 file from Hybrid Analysis.
      operationId: report_summary_api_v1_hybridanalysis_report_summary_post
      parameters:
      - in: query
        name: sha256
        required: true
        schema:
          title: Sha256
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Report Summary Api V1 Hybridanalysis Report Summary
                  Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Report Summary
  /api/v1/hybridanalysis/search-hash:
    post:
      description: Query hash value against Hybrid Analysis.
      operationId: search_hash_api_v1_hybridanalysis_search_hash_post
      parameters:
      - in: query
        name: hash_value
        required: true
        schema:
          title: Hash Value
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Search Hash Api V1 Hybridanalysis Search Hash Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Search Hash
...